Added Commands: View Selection Parse Nodes in Tooltip, View Selection JS in Tooltip
View nodes in tooltip displays the parse nodes generated from the selected JS and is assigned to ctrl+command+p by default.

View selection JS in tooltip displays the generated javascript in a tooltip and is assigned to ctrl+command+j by default.

Both are useful to make sure your CoffeeScript is doing what you think it should be doing.

Signed-off-by: Brandon Fryslie <sssssmokey@gmail.com>