Fixes #131. Change scope type to text.
It seems Sublime Text 2 have a habit of pushing subsequent lines of wrapped **code** one indentation level to the right. It does not do this with **text** (as opposed to code). Whether scope is considered code or text depends on `scopeName` property. If it starts with `source.` - then it's code. If it starts with `text.` - then it's text. Probably it would be better to make prose text and code - code, but I don't know how to.

I think this change makes sense in a way. In literate programming we focus on prose. So it's more text then source... Kinda :)